pyproject.toml
[tool.poetry]
name = "projektisininen"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
flask = "^2.2.2"
psycopg2-binary = "^2.9.5"
python-dotenv = "^0.21.0"
flask-sqlalchemy = "^3.0.2"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
